Add BandMatrix encode table to TQHD
Now it is possible to select the encode table used in TQHD. The table
used so far is now named as "BaseZero" since it starts from a word with
b bits of all 0s, and then flip one bit to 1per matrix row from left to
right.

The new matrix is named "BandMatrix" and is inspired by band matrices.
The idea is to create a diagonal band of 0s with top left and unter
right corners of 1s. Both matrices are valid thermometer codes since
each row is 1 bit away from each other (Hamming distance) and provide
the same accuracy.