Add accuracy dump when retraining
Add two new script arguments to retraining: "--retrain-dump-acc", and
"--retrain-dump-acc-suffix". The first argument activates the dumping
feature and allows the user to select the prefix path where the dumps
must be created. The scripts append a counter to indicate the iteration
being dumped. The user can use "--retrain-dump-acc-suffix" to complete
the path if necessary.