Centralize parallel job launch
Add a function to common.sh that is responsible for launching parallel
jobs. The function uses the "PROCFILE" feature of GNU parallel. Thus, a
temporary file is created in the file system containing the number of
jobs and the user may change the number of parallel jobs during script
execution.