Add HDCHOG
Add the HDCHOG pipeline based on the paper "HDCOG: A Lightweight
Hyperdimensional Computing Framework with Feature Extraction". The
script allows to choose between two HOG implementations. The first uses
skimage and the other is implemented on pytorch. The pytorch HOG is
faster than the skimage backend, but is not as flexible regarding
parameter customization.