Change emg accuracy dir from "emg" to "emg-all"
Store the accuracy of EMG all subjects in the folder "emg-all" instead
of "emg". The old style was "emg/hdc/all". The last level in the
hierarchy was different from the other apps and was always requiring
custom parsing in other scripts. By standardizing the hierarchy to
"emg-all/hdc", the results of EMG all subjects can easily be parsed.