Add TQHD normalization optimization
This commit adds a new feature to TQHD that allows it to optimize the
normalization step by caching the magnitude of a vector and reusing it
with future query vectors. When enabled, TQHD will cache the magnitude
of the first query vector computed. The reasoning is that all query
vectors in a model will have similar magnitudes. The option can be
activated with --am-tqhd-cache-norm in any application script.