Move unused deviation experiment with cache norm to script folder
The script paper-deviation-cache-norm.sh explores a possible
optimization of TQHD that caches the norm of a query vector and reuse
the value to speed reduce quantization processing. However, these
results are not present in the paper and this feature must be reworked.
Keep it in the repository for preservation purposes.