Add CacheDataset
Add a new feature that allows for caching a encoded dataset in HVs.
The cache is stored on disk and read on demand. This feature is helpful
when performing retraining or working with ultra large HVs. Disk
consumption may be high. The cache files are stored in a "_cache" folder.